requirements.txt
# ─── WEB FRAMEWORK ───────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.28.0
python-multipart==0.0.9
pydantic==2.6.0
pydantic-settings==2.2.0
httpx==0.27.0
python-jose[cryptography]==3.3.0

# ─── STREAMING ────────────────────────────────────────────────
kafka-python==2.0.2
aiokafka==0.10.0

# ─── ML / DEEP LEARNING CORE ──────────────────────────────────
torch==2.2.0
torchvision==0.17.0
tensorflow==2.15.0
transformers==4.38.0
datasets==2.18.0
accelerate==0.27.0
sentence-transformers==2.6.0
setfit==1.0.3
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
shap==0.44.1

# ─── GRAPH NEURAL NETWORKS ────────────────────────────────────
torch-geometric==2.5.0
networkx==3.2.1
neo4j==5.18.0

# ─── NLP SPECIFIC ─────────────────────────────────────────────
spacy==3.7.4
nltk==3.8.1
langdetect==1.0.9

# ─── COMPUTER VISION ──────────────────────────────────────────
opencv-python==4.9.0.80
Pillow==10.2.0
ultralytics==8.1.0
open-clip-torch==2.24.0
efficientnet-pytorch==0.7.1
imagehash==4.3.1
scikit-image==0.22.0
pytesseract==0.3.10

# ─── QR CODE ──────────────────────────────────────────────────
pyzxing==1.0.0
qrcode==7.4.2

# ─── STEGANOGRAPHY / FILE FORENSICS ───────────────────────────
python-magic==0.4.27
pyexiv2==2.9.0
exifread==3.0.0
stegano==0.11.2
lxml==5.1.0
scipy==1.12.0
numpy==1.26.4
yara-python==4.5.0
pywavelets==1.6.0

# ─── URL ANALYSIS ─────────────────────────────────────────────
tldextract==5.1.2
dnspython==2.6.1
python-whois==0.9.2
requests[security]==2.31.0
aiohttp==3.9.3
idna==3.6
confusable-homoglyphs==3.3.0

# ─── EMAIL ANALYSIS ───────────────────────────────────────────
mail-parser==3.15.0
dkimpy==1.1.5
checkdmarc==1.5.16
extract-msg==0.48.0

# ─── ONLINE LEARNING ──────────────────────────────────────────
river==0.21.0
evidently==0.4.22
mlflow==2.11.0

# ─── EDGE / ONNX ──────────────────────────────────────────────
onnx==1.16.0
onnxruntime==1.17.1
onnxruntime-gpu==1.17.1
optimum==1.18.0

# ─── DATABASES ────────────────────────────────────────────────
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.28
alembic==1.13.1
redis[hiredis]==5.0.2
motor==3.3.2
elasticsearch[async]==8.12.0
minio==7.2.5

# ─── MONITORING ───────────────────────────────────────────────
prometheus-client==0.20.0
structlog==24.1.0

# ─── THREAT INTEL CLIENTS ─────────────────────────────────────
vt-py==0.18.1
pymisp==2.4.185
OTXv2==1.5.12

# ─── TESTING ──────────────────────────────────────────────────
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
bandit==1.7.7
black==24.2.0
flake8==7.0.0
mypy==1.8.0
